server/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
pydantic
python-dotenv
psycopg2-binary
python-multipart
aiosmtplib